Refactor as lib/slack
