+++ routes/slack.js
@@ -2,7 +2,7 @@
 
 const express = require('express');
 const logger = require('heroku-logger');
-const controller = require('../controllers/slack');
+const slack = require('../lib/slack');
 
 const router = express.Router();
 
@@ -23,7 +23,7 @@ router.post('/', (req, res) => {
   const action = payload.actions[0];
 
   if (action.value === 'webSignup') {
-    return controller.sendSignup(channelId, userId, payload.callback_id);
+    return slack.sendSignup(channelId, userId, payload.callback_id);
   }
 
   logger.info('Unknown action', { action, userId, channelId });
